Pin libpng to at least 1.6.34
I am not entirely sure what's the issue here but 1.6.29 seems not to be
sufficient. We use 1.6.34 during the build, so let's require that version at
least. The conda-forge-pinnings pin to 1.6.34 anyway but I don't want to
rerender the 8.1 build again to drop some of the pins as I'd prefer to do that
on the 8.2 branch.

Fixes #33.